Update to 389 with Hadoop Support

+++ plugin/trino-accumulo/src/main/java/io/trino/plugin/accumulo/AccumuloClient.java
@@ -36,6 +36,7 @@ import io.trino.spi.TrinoException;
 import io.trino.spi.connector.ColumnMetadata;
 import io.trino.spi.connector.ConnectorSession;
 import io.trino.spi.connector.ConnectorTableMetadata;
+import io.trino.spi.connector.SchemaNotFoundException;
 import io.trino.spi.connector.SchemaTableName;
 import io.trino.spi.connector.TableNotFoundException;
 import io.trino.spi.predicate.Domain;
@@ -71,6 +72,7 @@ import java.util.stream.Collectors;
 import static io.trino.plugin.accumulo.AccumuloErrorCode.ACCUMULO_TABLE_DNE;
 import static io.trino.plugin.accumulo.AccumuloErrorCode.ACCUMULO_TABLE_EXISTS;
 import static io.trino.plugin.accumulo.AccumuloErrorCode.UNEXPECTED_ACCUMULO_ERROR;
+import static io.trino.plugin.accumulo.metadata.ZooKeeperMetadataManager.DEFAULT_SCHEMA;
 import static io.trino.spi.StandardErrorCode.ALREADY_EXISTS;
 import static io.trino.spi.StandardErrorCode.FUNCTION_IMPLEMENTATION_ERROR;
 import static io.trino.spi.StandardErrorCode.INVALID_TABLE_PROPERTY;
@@ -111,6 +113,23 @@ public class AccumuloClient
         this.indexLookup = requireNonNull(indexLookup, "indexLookup is null");
 
         this.auths = connector.securityOperations().getUserAuthorizations(username);
+
+        // The default namespace is created in ZooKeeperMetadataManager's constructor
+        if (!tableManager.namespaceExists(DEFAULT_SCHEMA)) {
+            tableManager.createNamespace(DEFAULT_SCHEMA);
+        }
+    }
+
+    public void createSchema(String schemaName)
+    {
+        metaManager.createSchema(schemaName);
+        tableManager.createNamespace(schemaName);
+    }
+
+    public void dropSchema(String schemaName)
+    {
+        metaManager.dropSchema(schemaName);
+        tableManager.dropNamespace(schemaName);
     }
 
     public AccumuloTable createTable(ConnectorTableMetadata meta)
@@ -134,11 +153,13 @@ public class AccumuloClient
                 AccumuloTableProperties.getSerializerClass(tableProperties),
                 AccumuloTableProperties.getScanAuthorizations(tableProperties));
 
+        // Make sure the namespace exists
+        if (!tableManager.namespaceExists(table.getSchema())) {
+            throw new SchemaNotFoundException(table.getSchema());
+        }
+
         // First, create the metadata
         metaManager.createTableMetadata(table);
-
-        // Make sure the namespace exists
-        tableManager.ensureNamespace(table.getSchema());
 
         // Create the Accumulo table if it does not exist (for 'external' table)
         if (!tableManager.exists(table.getFullTableName())) {
@@ -363,8 +384,8 @@ public class AccumuloClient
             localityGroupsBuilder.put(g.getKey(), familyBuilder.build());
         }
 
-        Map<String, Set<Text>> localityGroups = localityGroupsBuilder.build();
-        LOG.debug("Setting locality groups: {}", localityGroups);
+        Map<String, Set<Text>> localityGroups = localityGroupsBuilder.buildOrThrow();
+        LOG.debug("Setting locality groups: %s", localityGroups);
         tableManager.setLocalityGroups(table.getFullTableName(), localityGroups);
     }
 
@@ -543,6 +564,10 @@ public class AccumuloClient
 
     public void createView(SchemaTableName viewName, String viewData)
     {
+        if (!tableManager.namespaceExists(viewName.getSchemaName())) {
+            throw new SchemaNotFoundException(viewName.getSchemaName());
+        }
+
         if (getSchemaNames().contains(viewName.getSchemaName())) {
             if (getViewNames(viewName.getSchemaName()).contains(viewName.getTableName())) {
                 throw new TrinoException(ALREADY_EXISTS, "View already exists");
@@ -558,6 +583,10 @@ public class AccumuloClient
 
     public void createOrReplaceView(SchemaTableName viewName, String viewData)
     {
+        if (!tableManager.namespaceExists(viewName.getSchemaName())) {
+            throw new SchemaNotFoundException(viewName.getSchemaName());
+        }
+
         if (getView(viewName) != null) {
             metaManager.deleteViewMetadata(viewName);
         }
@@ -851,7 +880,7 @@ public class AccumuloClient
             // Swallow this exception so the query does not fail due to being unable
             // to locate the tablet server for the provided Key.
             // This is purely an optimization, but we will want to log the error.
-            LOG.error("Failed to get tablet location, returning dummy location", e);
+            LOG.error(e, "Failed to get tablet location, returning dummy location");
             return Optional.empty();
         }
     }
@@ -882,7 +911,7 @@ public class AccumuloClient
         catch (Exception e) {
             // Swallow this exception so the query does not fail due to being unable to locate the tablet server for the default tablet.
             // This is purely an optimization, but we will want to log the error.
-            LOG.error("Failed to get tablet location, returning dummy location", e);
+            LOG.error(e, "Failed to get tablet location, returning dummy location");
             return Optional.empty();
         }
     }
